Fix Windows/MSVC C4244 narrowing warnings: float/double to int
Add static_cast<> around float/double values narrowed to an
integer-family type (C4244), across dali-ui-foundation and samples.
Left size_t-alias-to-int narrowing untouched, since that is an
int<->int conversion needing separate case-by-case review.

+++ dali-ui-foundation/integration-api/label-impl.cpp
@@ -923,7 +923,7 @@ int LabelImpl::GetMarqueeGap() const
   {
     return mTextScroller->GetGap();
   }
-  return UiConfig::GetCurrent().GetMarqueeGap();
+  return static_cast<int>(UiConfig::GetCurrent().GetMarqueeGap());
 }
 
 void LabelImpl::SetMarqueeOrientation(Text::MarqueeOrientation orientation)
